orderms/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
azure-eventhub==5.11.5